hashing out ideas
